# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "OperationKind": {
        "enum": [
          "retention_cleanup"
        ],
        "title": "OperationKind",
        "type": "string"
      },
      "OperationRun": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "operation": {
            "$ref": "#/components/schemas/OperationKind"
          },
          "origin": {
            "$ref": "#/components/schemas/TriggerOrigin"
          },
          "requested_by": {
            "minLength": 1,
            "title": "Requested By",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RetentionCleanupReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "schedule_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Id"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/OperationRunStatus"
          },
          "trace_id": {
            "format": "uuid",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "operation",
          "origin",
          "requested_by",
          "status",
          "started_at"
        ],
        "title": "OperationRun",
        "type": "object"
      },
      "OperationRunStatus": {
        "enum": [
          "running",
          "succeeded",
          "failed"
        ],
        "title": "OperationRunStatus",
        "type": "string"
      },
...
      "ProblemDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "story_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Story Id"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "trace_id"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "RetentionCleanupCommand": {
        "additionalProperties": false,
        "properties": {
          "dry_run": {
            "default": true,
            "title": "Dry Run",
            "type": "boolean"
          },
          "operation": {
            "const": "retention_cleanup",
            "default": "retention_cleanup",
            "title": "Operation",
            "type": "string"
          },
          "requested_by": {
            "minLength": 1,
            "title": "Requested By",
            "type": "string"
          }
        },
        "required": [
          "requested_by"
        ],
        "title": "RetentionCleanupCommand",
        "type": "object"
      },
      "RetentionCleanupReport": {
        "additionalProperties": false,
        "properties": {
          "deleted_count": {
            "minimum": 0.0,
            "title": "Deleted Count",
            "type": "integer"
          },
          "dry_run": {
            "title": "Dry Run",
            "type": "boolean"
          },
          "eligible_count": {
            "minimum": 0.0,
            "title": "Eligible Count",
            "type": "integer"
          },
          "failed_count": {
            "minimum": 0.0,
            "title": "Failed Count",
            "type": "integer"
          },
          "finished_at": {
            "format": "date-time",
            "title": "Finished At",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RetentionItemResult"
            },
            "title": "Items",
            "type": "array"
          },
          "reclaimed_bytes": {
            "minimum": 0.0,
            "title": "Reclaimed Bytes",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          }
        },
        "required": [
          "dry_run",
          "started_at",
          "finished_at",
          "eligible_count",
          "deleted_count",
          "failed_count",
          "reclaimed_bytes",
          "items"
        ],
        "title": "RetentionCleanupReport",
        "type": "object"
      },
...
      "RoleKind": {
        "enum": [
          "narrator",
          "host"
        ],
        "title": "RoleKind",
        "type": "string"
      },
      "RoleProfile": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "default_emotion": {
            "default": "neutral",
            "minLength": 1,
            "title": "Default Emotion",
            "type": "string"
          },
          "default_pitch": {
            "default": 0.0,
            "maximum": 12.0,
            "minimum": -12.0,
            "title": "Default Pitch",
            "type": "number"
          },
          "default_speed": {
            "default": 1.0,
            "maximum": 1.65,
            "minimum": 0.6,
            "title": "Default Speed",
            "type": "number"
          },
          "display_name": {
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "gpt_weights_path": {
            "anyOf": [
              {
                "maxLength": 2048,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpt Weights Path"
          },
          "kind": {
            "$ref": "#/components/schemas/RoleKind"
          },
          "profile_id": {
            "format": "uuid",
            "title": "Profile Id",
            "type": "string"
          },
          "slug": {
            "maxLength": 64,
            "minLength": 2,
            "pattern": "^[a-z0-9][a-z0-9_-]*[a-z0-9]$",
            "title": "Slug",
            "type": "string"
          },
          "sovits_weights_path": {
            "anyOf": [
              {
                "maxLength": 2048,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sovits Weights Path"
          },
          "speaker_id": {
            "minLength": 1,
            "title": "Speaker Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "default": 1,
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          },
          "visual_assets": {
            "$ref": "#/components/schemas/RoleVisualAssets"
          }
        },
        "required": [
          "slug",
          "display_name",
          "kind",
          "speaker_id"
        ],
        "title": "RoleProfile",
        "type": "object"
      },
      "RoleProfileCreate": {
        "additionalProperties": false,
        "properties": {
          "default_emotion": {
            "default": "neutral",
            "minLength": 1,
            "title": "Default Emotion",
            "type": "string"
          },
          "default_pitch": {
            "default": 0.0,
            "maximum": 12.0,
            "minimum": -12.0,
            "title": "Default Pitch",
            "type": "number"
          },
          "default_speed": {
            "default": 1.0,
            "maximum": 1.65,
            "minimum": 0.6,
            "title": "Default Speed",
            "type": "number"
          },
          "display_name": {
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "gpt_weights_path": {
            "anyOf": [
              {
                "maxLength": 2048,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpt Weights Path"
          },
          "kind": {
            "$ref": "#/components/schemas/RoleKind"
          },
          "slug": {
            "maxLength": 64,
            "minLength": 2,
            "pattern": "^[a-z0-9][a-z0-9_-]*[a-z0-9]$",
            "title": "Slug",
            "type": "string"
          },
          "sovits_weights_path": {
            "anyOf": [
              {
                "maxLength": 2048,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sovits Weights Path"
          },
          "speaker_id": {
            "minLength": 1,
            "title": "Speaker Id",
            "type": "string"
          },
          "visual_assets": {
            "$ref": "#/components/schemas/RoleVisualAssets"
          }
        },
        "required": [
          "slug",
          "display_name",
          "kind",
          "speaker_id"
        ],
        "title": "RoleProfileCreate",
        "type": "object"
      },
      "RoleProfileDelete": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "RoleProfileDelete",
        "type": "object"
      },
      "RoleProfileReplace": {
        "additionalProperties": false,
        "properties": {
          "default_emotion": {
            "default": "neutral",
            "minLength": 1,
            "title": "Default Emotion",
            "type": "string"
          },
          "default_pitch": {
            "default": 0.0,
            "maximum": 12.0,
            "minimum": -12.0,
            "title": "Default Pitch",
            "type": "number"
          },
          "default_speed": {
            "default": 1.0,
            "maximum": 1.65,
            "minimum": 0.6,
            "title": "Default Speed",
            "type": "number"
          },
          "display_name": {
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "gpt_weights_path": {
            "anyOf": [
              {
                "maxLength": 2048,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpt Weights Path"
          },
          "kind": {
            "$ref": "#/components/schemas/RoleKind"
          },
          "slug": {
            "maxLength": 64,
            "minLength": 2,
            "pattern": "^[a-z0-9][a-z0-9_-]*[a-z0-9]$",
            "title": "Slug",
            "type": "string"
          },
          "sovits_weights_path": {
            "anyOf": [
              {
                "maxLength": 2048,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sovits Weights Path"
          },
          "speaker_id": {
            "minLength": 1,
            "title": "Speaker Id",
            "type": "string"
          },
          "visual_assets": {
            "$ref": "#/components/schemas/RoleVisualAssets"
          }
        },
        "required": [
          "slug",
          "display_name",
          "kind",
          "speaker_id",
          "expected_version"
        ],
        "title": "RoleProfileReplace",
        "type": "object"
      },
      "RoleVisualAssets": {
        "additionalProperties": false,
        "description": "Renderer-neutral character asset references.\n\nReferences are identifiers only. Resolving and reading an asset belongs to a\nfuture renderer adapter, so the profile service never guesses a filesystem\nor URL contract.",
        "properties": {
          "differential_art": {
            "items": {
              "$ref": "#/components/schemas/DifferentialArtAsset"
            },
            "maxItems": 100,
            "title": "Differential Art",
            "type": "array"
          },
          "live2d_asset_ref": {
            "anyOf": [
              {
                "maxLength": 2048,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Live2D Asset Ref"
          }
        },
        "title": "RoleVisualAssets",
        "type": "object"
      },
      "ScheduleSnapshot": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "interval_seconds": {
            "exclusiveMinimum": 0.0,
            "title": "Interval Seconds",
            "type": "number"
          },
          "last_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run Id"
          },
          "last_run_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OperationRunStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "next_run_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Run At"
          },
          "operation": {
            "$ref": "#/components/schemas/OperationKind"
          },
          "schedule_id": {
            "minLength": 1,
            "title": "Schedule Id",
            "type": "string"
          }
        },
        "required": [
          "schedule_id",
          "operation",
          "enabled",
          "interval_seconds"
        ],
        "title": "ScheduleSnapshot",
        "type": "object"
      },
...
      "TriggerOrigin": {
        "enum": [
          "manual",
          "schedule"
        ],
        "title": "TriggerOrigin",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Review-gated multilingual content-to-audio pipeline.",
    "title": "god-news",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/operations/retention/runs": {
      "post": {
        "operationId": "run_retention_api_v1_operations_retention_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetentionCleanupCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Run Retention"
      }
    },
    "/api/v1/operations/runs": {
      "get": {
        "operationId": "list_operation_runs_api_v1_operations_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OperationRun"
                  },
                  "title": "Response List Operation Runs Api V1 Operations Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Operation Runs"
      }
    },
    "/api/v1/operations/schedules": {
      "get": {
        "operationId": "list_operation_schedules_api_v1_operations_schedules_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScheduleSnapshot"
                  },
                  "title": "Response List Operation Schedules Api V1 Operations Schedules Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Operation Schedules"
      }
    },
    "/api/v1/roles": {
      "get": {
        "operationId": "list_roles_api_v1_roles_get",
        "parameters": [
          {
            "in": "query",
            "name": "enabled",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Enabled"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoleProfile"
                  },
                  "title": "Response List Roles Api V1 Roles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Roles"
      },
      "post": {
        "operationId": "create_role_api_v1_roles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleProfileCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Role"
      }
    },
    "/api/v1/roles/{profile_id}": {
      "delete": {
        "operationId": "delete_role_api_v1_roles__profile_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleProfileDelete"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Delete Role"
      },
      "get": {
        "operationId": "get_role_api_v1_roles__profile_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Role"
      },
      "put": {
        "operationId": "replace_role_api_v1_roles__profile_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleProfileReplace"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Replace Role"
      }
    },
...
  }
}